refactor: split utilities to own files
To avoid the API class becoming a monolith containing every function we
can split parts out to their own files and classes.

The `euclideanDistance` export is not strictly necessary but is an
example of how a function from another file can be added to the API.
