+++ scripts/api.js
@@ -1,5 +1,10 @@
 import {getMessageInfo} from "./messageParser.js";
+import {euclideanDistance} from "./utils.js";
+import LloydsAlgorithm from "./lloydsAlgorithm.js";
 
+/**
+ * Functions exposed by the module for use in macros.
+ */
 class ModuleApi {
     static getEffectVolume(volume) {
         return volume * game.settings.get("lancer-weapon-fx", "volume");
@@ -28,76 +33,15 @@ class ModuleApi {
         };
     }
 
-    static euclideanDistance (point1, point2) {
-        // Calculate the Euclidean distance between two points.
-        const dx = point1.x - point2.x;
-        const dy = point1.y - point2.y;
-        return Math.sqrt(dx * dx + dy * dy);
-    }
-
-    static assignToClusters (points, centroids) {
-        const clusters = new Array(centroids.length).fill(null).map(() => []);
-
-        for (const point of points) {
-            const distances = centroids.map(centroid => this.euclideanDistance(point, centroid));
-            const clusterIndex = distances.indexOf(Math.min(...distances));
-            clusters[clusterIndex].push(point);
-        }
-
-        return clusters;
-    }
-
-    static calculateCentroids (cluster) {
-        // Calculate the centroid (mean) of a cluster of points.
-        if (cluster.length === 0) {
-          return null;
-        }
-
-        const sumX = cluster.reduce((sum, point) => sum + point.x, 0);
-        const sumY = cluster.reduce((sum, point) => sum + point.y, 0);
-
-        return { x: sumX / cluster.length, y: sumY / cluster.length };
-    }
-
-    static lloydsAlgorithm (points, numCentroids) {
-        if (numCentroids <= 0) {
-          return [];
-        }
-
-        // Initialize random centroids to start the algorithm.
-        const centroids = [];
-        for (let i = 0; i < numCentroids; i++) {
-          centroids.push(points[Math.floor(Math.random() * points.length)]);
-        }
-
-        let previousCentroids;
-        while (
-          !previousCentroids ||
-          !centroids.every((centroid, i) =>
-            centroid.x === previousCentroids[i].x && centroid.y === previousCentroids[i].y
-          )
-        ) {
-          previousCentroids = centroids.map(centroid => ({ ...centroid }));
-
-          // Assign points to clusters based on the current centroids.
-          const clusters = this.assignToClusters(points, centroids);
-
-          // Calculate new centroids based on the current cluster assignments.
-          centroids.forEach((_, i) => {
-            centroids[i] = this.calculateCentroids(clusters[i]);
-          });
-        }
-
-        return centroids;
-    }
-
     static getTargetLocationsFromTokenGroup(targetTokens, numGroups) {
         const targetPoints = targetTokens.map(token => {
             return {x: token.center.x, y: token.center.y};
         });
 
-        return this.lloydsAlgorithm(targetPoints, numGroups);
+        return LloydsAlgorithm.getCentroids(targetPoints, numGroups);
     }
+
+    static euclideanDistance = euclideanDistance;
 }
 
 Hooks.on("init", () => game.modules.get("lancer-weapon-fx").api = ModuleApi);
